add permissions

+++ app/Http/Controllers/Auth/RegisteredUserController.php
@@ -11,6 +11,7 @@ use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rules;
 use Illuminate\View\View;
@@ -71,9 +72,8 @@ class RegisteredUserController extends Controller
     }
     public function destroy(User $user)
     {
+        Gate::authorize('delete',Auth::user());
         $user->delete();
-
-
        return redirect()->route('users.index')->with('success', "Book $user->name delete successfully");
     }
